DjangoAPI/requirements.txt
Django==2.2
gunicorn==20.0.4
django-cors-headers
djangorestframework==3.9.2

